refact : Router 사용하여 Loading/Home 페이지 분리
- 기존 App 컴포넌트에서 data fetching을 하던 것에서 React-Router를 사용하여 Loading 페이지에서 data fetching 후 Home으로 이동하게 바꿈
- AddLocation 페이지는 UI변경에 따라 불 필요하여 삭제함
