+++ src/App.js
@@ -1,48 +1,18 @@
-import React, { useState, useEffect, useCallback } from 'react';
+import React from 'react';
 import styled from 'styled-components';
-import { API_KEY } from './config';
 import Loading from './pages/Loading';
 import Home from './pages/Home';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 const App = () => {
-	const [loading, setLoading] = useState(true);
-	const [result, setResult] = useState({});
-
-	// 현재 위치 가져오기
-	const getCurrentLocation = useCallback(() => {
-		navigator.geolocation.getCurrentPosition((position) => {
-			const lat = position.coords.latitude; // 위도
-			const lon = position.coords.longitude; // 경도
-
-			getCurrentWeather(lat, lon);
-		});
-	}, []);
-
-	// 현재 날씨 가져오기
-	const getCurrentWeather = async (lat, lon) => {
-		const url = `https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=${API_KEY}&units=metric&lang=kr`;
-
-		try {
-			const result = await fetch(url).then((response) => response.json());
-			console.log(result);
-			setResult(result);
-			setLoading(false); // 날씨 가져오기 성공하면 로딩 false
-		} catch (error) {
-			console.error(error);
-		}
-	};
-
-	// 앱 마운트 시,
-	useEffect(() => {
-		getCurrentLocation();
-		return () => {
-			console.log('app unmounted');
-		};
-	}, [getCurrentLocation]);
-
 	return (
 		<AppContainer className="App">
-			{loading ? <Loading /> : <Home weather={result} />}
+			<BrowserRouter>
+				<Routes>
+					<Route path="/" element={<Loading />} />
+					<Route path="/home" element={<Home />} />
+				</Routes>
+			</BrowserRouter>
 		</AppContainer>
 	);
 };
