--- src/pages/AddLocation.js
@@ -1,8 +1,0 @@
-// 주소 찾기 페이지 - 검색창, 현재 위치 찾기 버튼, 검색결과
-import React from 'react';
-
-const AddLocation = () => {
-	return <div></div>;
-};
-
-export default AddLocation;
