+++ src/pages/Home.js
@@ -8,10 +8,15 @@ import WeatherContainer from '../components/weather/WeatherContainer';
 import SideBar from '../components/sideBar/SideBar';
 import { registerDragEvent } from '../utils/drag';
 import { inrange } from './../utils/drag';
+import { useLocation } from 'react-router-dom';
 
-const Home = ({ weather }) => {
+const Home = () => {
 	const [modalVisible, setModalVisible] = useState(false);
 	const [dragUp, setDragUp] = useState(0);
+
+	// 날씨 정보 받아오기
+	const location = useLocation();
+	const WEATHER = location.state.result;
 
 	const openModal = () => {
 		setModalVisible(true);
@@ -23,7 +28,7 @@ const Home = ({ weather }) => {
 
 	return (
 		<HomeContainer>
-			<Header location={weather.name} openModal={openModal} />
+			<Header location={WEATHER.name} openModal={openModal} />
 			<main
 				// dragUp 이벤트
 				{...registerDragEvent({
@@ -40,8 +45,8 @@ const Home = ({ weather }) => {
 					},
 				})}
 			>
-				<Dresses temperature={Math.round(weather.main.temp)} />
-				<WeatherContainer weather={weather} dragUp={dragUp} />
+				<Dresses temperature={Math.round(WEATHER.main.temp)} />
+				<WeatherContainer weather={WEATHER} dragUp={dragUp} />
 			</main>
 			{modalVisible && <SideBar onClose={closeModal} />}
 		</HomeContainer>
